requirements.txt
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
crispy-bootstrap4==2025.6
cryptography==46.0.3
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2.7
django-allauth==65.12.1
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.15.0
djangorestframework-stubs==3.16.5
djangorestframework-xml==2.0.0
flake8==6.1.0
freetype-py==2.5.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
idna==3.11
lxml==6.0.2
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
mysqlclient==2.2.7
oauthlib==3.3.1
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
playwright==1.55.0
psycopg==3.2.11
psycopg-binary==3.2.11
psycopg2-binary==2.9.11
pycairo==1.28.0
pycodestyle==2.11.1
pycparser==2.23
pyee==13.0.0
pyflakes==3.1.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pypdf==6.1.3
PySocks==1.7.1
python-bidi==0.6.7
python-dotenv==1.2.1
PyYAML==6.0.3
reportlab==4.4.4
requests==2.32.4
requests-oauthlib==1.3.1
responses==0.25.8
rlPyCairo==0.4.0
ruff==0.1.0
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
trio==0.32.0
trio-websocket==0.12.2
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
wsproto==1.2.0
xhtml2pdf==0.2.17